Add API function for using cleanly as a library #39
