+++ setup.cfg
@@ -86,6 +86,7 @@ testing =
     black
     isort
     pycodestyle
+    pytest-rerunfailures
 
 docs =
     Sphinx >= 3.3.1
