+++ tests/data/azure-devops.req-310-expected.json
@@ -2,7 +2,7 @@
   "headers": {
     "tool_name": "python-inspector",
     "tool_homepageurl": "https://github.com/nexB/python-inspector",
-    "tool_version": "0.8.5",
+    "tool_version": "0.9.0",
     "options": [
       "--requirement /home/tg1999/Desktop/python-inspector-1/tests/data/azure-devops.req.txt",
       "--index-url https://pypi.org/simple",
@@ -14,6 +14,109 @@
     "warnings": [],
     "errors": []
   },
+  "resolved_dependencies_graph": [
+    {
+      "package": "pkg:pypi/azure-core@1.26.0",
+      "dependencies": [
+        "pkg:pypi/requests@2.28.1",
+        "pkg:pypi/six@1.16.0",
+        "pkg:pypi/typing-extensions@4.4.0"
+      ]
+    },
+    {
+      "package": "pkg:pypi/azure-devops@6.0.0b4",
+      "dependencies": [
+        "pkg:pypi/msrest@0.6.21"
+      ]
+    },
+    {
+      "package": "pkg:pypi/azure-storage-blob@12.13.1",
+      "dependencies": [
+        "pkg:pypi/azure-core@1.26.0",
+        "pkg:pypi/cryptography@38.0.2",
+        "pkg:pypi/msrest@0.6.21"
+      ]
+    },
+    {
+      "package": "pkg:pypi/certifi@2022.9.24",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/cffi@1.15.1",
+      "dependencies": [
+        "pkg:pypi/pycparser@2.21"
+      ]
+    },
+    {
+      "package": "pkg:pypi/charset-normalizer@2.1.1",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/click@8.1.3",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/cryptography@38.0.2",
+      "dependencies": [
+        "pkg:pypi/cffi@1.15.1"
+      ]
+    },
+    {
+      "package": "pkg:pypi/idna@3.4",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/isodate@0.6.1",
+      "dependencies": [
+        "pkg:pypi/six@1.16.0"
+      ]
+    },
+    {
+      "package": "pkg:pypi/msrest@0.6.21",
+      "dependencies": [
+        "pkg:pypi/certifi@2022.9.24",
+        "pkg:pypi/isodate@0.6.1",
+        "pkg:pypi/requests-oauthlib@1.3.1",
+        "pkg:pypi/requests@2.28.1"
+      ]
+    },
+    {
+      "package": "pkg:pypi/oauthlib@3.2.2",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/pycparser@2.21",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/requests-oauthlib@1.3.1",
+      "dependencies": [
+        "pkg:pypi/oauthlib@3.2.2",
+        "pkg:pypi/requests@2.28.1"
+      ]
+    },
+    {
+      "package": "pkg:pypi/requests@2.28.1",
+      "dependencies": [
+        "pkg:pypi/certifi@2022.9.24",
+        "pkg:pypi/charset-normalizer@2.1.1",
+        "pkg:pypi/idna@3.4",
+        "pkg:pypi/urllib3@1.26.12"
+      ]
+    },
+    {
+      "package": "pkg:pypi/six@1.16.0",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/typing-extensions@4.4.0",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/urllib3@1.26.12",
+      "dependencies": []
+    }
+  ],
   "files": [
     {
       "type": "file",
@@ -120,109 +223,6 @@
           "purl": null
         }
       ]
-    }
-  ],
-  "resolved_dependencies_graph": [
-    {
-      "package": "pkg:pypi/azure-core@1.26.0",
-      "dependencies": [
-        "pkg:pypi/requests@2.28.1",
-        "pkg:pypi/six@1.16.0",
-        "pkg:pypi/typing-extensions@4.4.0"
-      ]
-    },
-    {
-      "package": "pkg:pypi/azure-devops@6.0.0b4",
-      "dependencies": [
-        "pkg:pypi/msrest@0.6.21"
-      ]
-    },
-    {
-      "package": "pkg:pypi/azure-storage-blob@12.13.1",
-      "dependencies": [
-        "pkg:pypi/azure-core@1.26.0",
-        "pkg:pypi/cryptography@38.0.2",
-        "pkg:pypi/msrest@0.6.21"
-      ]
-    },
-    {
-      "package": "pkg:pypi/certifi@2022.9.24",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/cffi@1.15.1",
-      "dependencies": [
-        "pkg:pypi/pycparser@2.21"
-      ]
-    },
-    {
-      "package": "pkg:pypi/charset-normalizer@2.1.1",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/click@8.1.3",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/cryptography@38.0.2",
-      "dependencies": [
-        "pkg:pypi/cffi@1.15.1"
-      ]
-    },
-    {
-      "package": "pkg:pypi/idna@3.4",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/isodate@0.6.1",
-      "dependencies": [
-        "pkg:pypi/six@1.16.0"
-      ]
-    },
-    {
-      "package": "pkg:pypi/msrest@0.6.21",
-      "dependencies": [
-        "pkg:pypi/certifi@2022.9.24",
-        "pkg:pypi/isodate@0.6.1",
-        "pkg:pypi/requests-oauthlib@1.3.1",
-        "pkg:pypi/requests@2.28.1"
-      ]
-    },
-    {
-      "package": "pkg:pypi/oauthlib@3.2.2",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/pycparser@2.21",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/requests-oauthlib@1.3.1",
-      "dependencies": [
-        "pkg:pypi/oauthlib@3.2.2",
-        "pkg:pypi/requests@2.28.1"
-      ]
-    },
-    {
-      "package": "pkg:pypi/requests@2.28.1",
-      "dependencies": [
-        "pkg:pypi/certifi@2022.9.24",
-        "pkg:pypi/charset-normalizer@2.1.1",
-        "pkg:pypi/idna@3.4",
-        "pkg:pypi/urllib3@1.26.12"
-      ]
-    },
-    {
-      "package": "pkg:pypi/six@1.16.0",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/typing-extensions@4.4.0",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/urllib3@1.26.12",
-      "dependencies": []
     }
   ],
   "packages": [
