+++ tests/data/insecure-setup/setup.py-expected.json
@@ -2,7 +2,7 @@
   "headers": {
     "tool_name": "python-inspector",
     "tool_homepageurl": "https://github.com/nexB/python-inspector",
-    "tool_version": "0.8.5",
+    "tool_version": "0.9.0",
     "options": [
       "--index-url https://pypi.org/simple",
       "--python-version 27",
@@ -13,6 +13,22 @@
     "warnings": [],
     "errors": []
   },
+  "resolved_dependencies_graph": [
+    {
+      "package": "pkg:pypi/isodate@0.6.1",
+      "dependencies": [
+        "pkg:pypi/six@1.16.0"
+      ]
+    },
+    {
+      "package": "pkg:pypi/pyparsing@2.4.7",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/six@1.16.0",
+      "dependencies": []
+    }
+  ],
   "files": [
     {
       "type": "file",
@@ -118,22 +134,6 @@
           "purl": "pkg:pypi/rdflib@5.0.0"
         }
       ]
-    }
-  ],
-  "resolved_dependencies_graph": [
-    {
-      "package": "pkg:pypi/isodate@0.6.1",
-      "dependencies": [
-        "pkg:pypi/six@1.16.0"
-      ]
-    },
-    {
-      "package": "pkg:pypi/pyparsing@2.4.7",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/six@1.16.0",
-      "dependencies": []
     }
   ],
   "packages": [
