+++ tests/data/setup/simple-setup.py-expected.json
@@ -2,7 +2,7 @@
   "headers": {
     "tool_name": "python-inspector",
     "tool_homepageurl": "https://github.com/nexB/python-inspector",
-    "tool_version": "0.8.5",
+    "tool_version": "0.9.0",
     "options": [
       "--index-url https://pypi.org/simple",
       "--python-version 27",
@@ -13,6 +13,18 @@
     "warnings": [],
     "errors": []
   },
+  "resolved_dependencies_graph": [
+    {
+      "package": "pkg:pypi/boolean-py@3.8",
+      "dependencies": []
+    },
+    {
+      "package": "pkg:pypi/license-expression@1.0",
+      "dependencies": [
+        "pkg:pypi/boolean-py@3.8"
+      ]
+    }
+  ],
   "files": [
     {
       "type": "file",
@@ -74,18 +86,6 @@
           "datasource_id": "pypi_setup_py",
           "purl": "pkg:pypi/example-app@2.4.0"
         }
-      ]
-    }
-  ],
-  "resolved_dependencies_graph": [
-    {
-      "package": "pkg:pypi/boolean-py@3.8",
-      "dependencies": []
-    },
-    {
-      "package": "pkg:pypi/license-expression@1.0",
-      "dependencies": [
-        "pkg:pypi/boolean-py@3.8"
       ]
     }
   ],
